// src/lib/client/models/FilesystemLocation.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FilesystemLocation = {
    id?: (number | null);
    key?: (string | null);
    label: string;
    name: string;
    path: string;
    type?: string;
    description?: (string | null);
    is_editable?: boolean;
    created_at?: (string | null);
    updated_at?: (string | null);
    exists?: (boolean | null);
};